Fix mypy typechecking
Two major changes here:

1. Pull in the PyActionNetwork code so we can manage it directly. It doesn't have a
`py.typed` so mypy doesn't know what to do with it. Ideally, we'll expand this, type
it more properly, then extract it to its own package, but this will work for now.
2. Switch to pyairtable. This is also properly typed plus has an ORM that we may want
to take advantage of later. This is also the updated version of
airtable-python-wrapper, so will be better supported going forward.